@@ -4,6 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "dev": "./node_modules/.bin/webpack-dev-server --progress --colors",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -11,20 +12,29 @@
   "dependencies": {
     "babel-core": "^6.5.2",
     "babel-loader": "^6.2.2",
+    "babel-plugin-react-html-attrs": "^2.0.0",
+    "babel-plugin-transform-class-properties": "^6.3.13",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
+    "bootstrap-without-jquery": "^1.0.4",
+    "dotenv": "^2.0.0",
+    "flux": "^2.1.1",
     "install": "^0.8.1",
     "lodash": "^4.15.0",
     "npm": "^3.10.6",
-    "react": "^0.14.7",
+    "react": "^15.3.1",
     "react-dom": "^15.3.1",
-    "react-toolbox": "^1.1.2"
+    "react-router": "^2.7.0",
+    "react-toolbox": "^1.1.2",
+    "source-map-loader": "^0.1.5"
   },
   "devDependencies": {
     "node-sass": "^3.8.0",
     "sass-loader": "^4.0.0",
     "source-map-loader": "^0.1.5",
-    "webpack": "^1.13.2"
+    "webpack": "^1.13.2",
+    "webpack-dev-server": "^1.15.0"
   }
 }
