@@ -1,4 +1,7 @@
-import React from "React";
+import React from "react";
+import * as BookingActions from "../actions/bookingActions";
+import BookingStore from "../stores/bookingStore";
+import { withRouter } from 'react-router'
 
 // Booking Module Structure
 //
@@ -9,41 +12,85 @@ import React from "React";
 //   |- BookingForm
 
 export class BookingComponent extends React.Component {
+  constructor(props){
+    super(props);
+    console.log(this.props)
+    console.log(withRouter);
+    setTimeout(()=> this.props.router.push('/about'), 1000);
+  
+    this.getDates = this.getDates.bind(this);
+    this.state = {
+      dates: []
+    };
+  }
+
+  componentWillMount() {
+    this.getDates();
+    BookingStore.on("change", this.getDates);
+  }
+
+  componentDidMount(){
+    BookingActions.reloadDates();
+  }
+
+  componentWillUnmount() {
+    BookingStore.removeListener("change", this.getDates);
+  }
+
+  getDates() {
+    this.setState({
+      dates: BookingStore.getAll(),
+    });
+  }
+
+  reloadDates() {
+    BookingActions.reloadDates();
+  }
+
   render() {
-    var data = [
-      {'id':1, 'user': 'Pepe', 'date': '2016-12-12'},
-      {'id':2, 'user': 'Marcos', 'date': '2016-11-12'},
-      //{'user': 'Andres', 'date': '2016-10-12'},
-      {'id':3, 'user': 'Andres'},
-      {'id':4, 'user': 'Hernan', 'date': '2016-9-12'}
-    ];
-    return <div>
-      <BookingList list={data} />
-      <BookingForm />
-    </div>
+    return (<div>
+      <BookingList dates={this.state.dates} />
+    </div>)
   }
 }
 // BookingList
 class BookingList extends React.Component {
   render() {
-  	var booking = this.props.list.map(function(book){
-  		return <BookingListItem key={book.id} user={book.user} date={book.date}/>
+  	var booking = this.props.dates.map(function(date, i){
+  		return <BookingListItem key={i} userId={date.userId} desc={date.desc} providerId={date.providerId}/>
   	})
 		
-    return <div>{booking}</div>
+    return <div>
+            <BookingForm />
+            <div>{booking}</div>
+          </div>
   }
 }
 
 // BookingListItem
 class BookingListItem extends React.Component {
   render() {
-    return <div> User: {this.props.user} <br/> date: {this.props.date}</div>
+    return (<div class="col-md-3 col-sm-6 hero-feature">
+              <div class="thumbnail">
+                  <img src="http://placehold.it/800x500" alt=""/>
+                  <div class="caption">
+                      <p>User: {this.props.userId}</p>
+                      <p>User: {this.props.providerId}</p>
+                      <p>{this.props.desc}</p>
+                      <p>
+                          <a href="#" class="btn btn-primary">Buy Now!</a>
+                          <a href="#" class="btn btn-default">More Info</a>
+                      </p>
+                  </div>
+              </div>
+          </div>)  
   }
 }
 
 BookingListItem.propTypes = {
-  user: React.PropTypes.string.isRequired,
-  date: React.PropTypes.string.isRequired
+  userId: React.PropTypes.number.isRequired,
+  providerId: React.PropTypes.number.isRequired,
+  desc: React.PropTypes.string.isRequired
 };
 
 BookingListItem.defaultProps = {
@@ -54,11 +101,6 @@ BookingListItem.defaultProps = {
 class BookingForm extends React.Component {
   constructor(props) {
     super(props);
-    // this.state.user = {
-    //   user: '',
-    //   date: ''
-    // };
-
     this.state = {
       user : Object.assign({})
     }
@@ -86,7 +128,6 @@ class BookingForm extends React.Component {
 
   submitForm(event){
     event.preventDefault();
-    console.log(this.state.user);
     this.setState({
       user : {}
     });
@@ -95,13 +136,34 @@ class BookingForm extends React.Component {
   render() {
     return <div>
       <h3>Reserve a date</h3>
-        <form action="">
-        User: <input value={this.state.user.user} type="text" name="user" onChange={this.handleChange}/>
-        <br/>
-        Date: <input value={this.state.user.date} type="text" name="date" onChange={this.handleChange}/>
-        <br/>
-        <input type="submit" onClick={this.submitForm.bind(this)} value="Send"/>
-      </form>
+        <div class="col-md-3 col-sm-6 hero-feature">
+            <div class="thumbnail">
+                <img src="http://placehold.it/250x80?text=Add%20New" alt=""/>
+                <div class="caption">
+                <div class="form-group">
+                  <div class="input-group">
+                    <div class="input-group-addon">User</div>
+                    <input class="form-control" value={this.state.user.user} type="text" name="user" placeholder="UserName" onChange={this.handleChange}/>
+                  </div>
+                </div>
+                <div class="form-group">
+                  <div class="input-group">
+                    <div class="input-group-addon">Provider</div>
+                    <input class="form-control" value={this.state.user.date} type="text" name="date" placeholder="Provider" onChange={this.handleChange}/>
+                  </div>
+                </div>
+                  <div class="form-group">
+                    <select class="form-control" name="service" id="">
+                      <option value="Complete">Complete</option>
+                      <option value="Complete">Complete + 1</option>
+                      <option value="Complete">Complete + 2</option>
+                    </select>
+                    <br/>
+                    <button type="submit" onClick={this.submitForm.bind(this)} class="btn btn-success">Add New</button>
+                </div>
+                </div>
+            </div>
+        </div>
     </div>
   }
 }
