@@ -1,12 +1,16 @@
-import React from "React";
-import { BookingComponent } from './booking';
-
-export interface ApplicationProps { }
+import React from "react";
+import { Router, Route, browserHistory } from 'react-router'
+import { BookingComponent } from './booking'
 
 export class Application extends React.Component {
     render() {
-        return <div>
-                <BookingComponent/>
-               </div>;
+        return (
+            <Router history={browserHistory}>
+                <Route path="/" component={BookingComponent}/>
+                <Route path="about" component={BookingComponent}>
+                    <Route path="me" component={BookingComponent}/>
+                </Route>
+            </Router>
+        );
     }
 }
