@@ -1,6 +1,5 @@
-import React from 'React';
-import ReactDOM from 'React-dom';
-
+import React from 'react';
+import ReactDOM from 'react-dom';
 import { Application } from "./components/application";
 
 ReactDOM.render(
