@@ -29,6 +29,11 @@ module.exports = {
                         'babel-preset-react',
                         'babel-preset-stage-0',
                       ].map(require.resolve),
+                    plugins: [
+                        'react-html-attrs',
+                        'transform-class-properties',
+                        'transform-decorators-legacy'
+                        ],
                   }
             },
             {
@@ -49,8 +54,8 @@ module.exports = {
     // assume a corresponding global variable exists and use that instead.
     // This is important because it allows us to avoid bundling all of our
     // dependencies, which allows browsers to cache those libraries between builds.
-    externals: {
-        "react": "React",
-        'react-dom': 'ReactDOM'
-    },
+    // externals: {
+    //     "react": "React",
+    //     'react-dom': 'ReactDOM'
+    // },
 };
